travis: drop cask usage for osxfuse
Whilst Travis continues to build on 10.9.x we're able to use Homebrew's
osxfuse for testing. This is preferable as it reduces the level of variables
we need to handle, and it means the bot won't fail with conflict messages as
it currently is.

Once Travis makes the jump to 10.10.x we'll need to go back to using the Cask.